Cruft removal part 3:
In scsynth's wavetable code (SC_Samp.h & Samp.cpp) there's "gPMSine" instantiated but never used. It seems it was intended for use with the "PMOsc" code which was never implemented as a real sc3 ugen. There's no code that uses it and it's not exposed in the plugin interface. Yet the table is created and filled every single time scsynth boots. Now removed.

The same files also define a WavetableAsSignal() method, which is never used by anything. (Note: there's a method used by the lang which has the same name, and this IS used, I'm not suggesting removing that.) Now removed.


git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@7831 a380766d-ff14-0410-b294-a243070f3f08