PyrSlot api cleanup:
this patch adds a level of abstraction to the PyrSlot access. instead
of directly accessing members using accessor macros, a number of getter
and setter inline functions are introduced. in future this should
simplify the potability to 64bit platforms, since only the getter/setter
api needs to be adapted.

this patch doesn't introduce any functional change.

Signed-off-by: Tim Blechmann <tim@klingt.org>

git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@9597 a380766d-ff14-0410-b294-a243070f3f08