supernova: introduce locking for buses and buffers as required for supernova
for supernova buses and buffers need to be guarded by reader/writer
spinlocks to ensure data consistency.

buffer locks are added as new member to the SndBuf class, bus locks are
stored in an array in the World class.

locks are acquired using a macro interface, that is enabled in the
SUPERNOVA preprocessor symbol is defined.

Signed-off-by: Tim Blechmann <tim@klingt.org>

git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@10161 a380766d-ff14-0410-b294-a243070f3f08