sync condition: move implementation to header file
SC_SyncCondition.h is used both in the server and in the plugins. the
header should therefore be placed in the `common' directory. the actual
implementation is also quite trivial, so it should be ok to inline it
and to completely remove the source file. this also simplifies the build
process.

Signed-off-by: Tim Blechmann <tim@klingt.org>