Pipe does not remove closed pipes from openFiles
Found this in 3.4 -- I don't know if it's fixed in master.

There's an object leak in Pipe. When you close the pipe, the Pipe object
remains in the UnixFILE.openFiles array. It's properly closed in the
backend -- fileptr is nil -- but the object will never get garbage
collected.

One fix is to replicate the way File does it. The other way (maybe better,
but it might have some unintended consequences I can't think of at the
moment) would be to promote close and prClose up to UnixFILE, and let File
and Pipe inherit those methods. That would at least reduce the risk of
somebody adding something clever to one and not adding it to the other.

Signed-off-by: Tim Blechmann <tim@klingt.org>