QtCollider: implement setting/getting global palette
QtGUI.palette and QtGUI.palette_ get and set the global Qt palette,
that represents default palette for all widgets.

After setting a new global palette, all the existing widgets will
update those colors that have not been changed manually to the new
defaults.
Newly created widgets will immediately use the new global palette.

If a color role in a QPalette SC object is set to nil, it refers to
the color for the same role from the global palette.