QHelpBrowser: a better fix for anchors in URL not having effect
QWebView makes use of anchor (fragment) in the given address only
if the address is already in the URL format, not the native filesystem
format ("file:///" has to be prepended).

If there is no scheme detected in the given address (as is the case
with native filesystem addresses), force the scheme "file".