SCDoc: show inherited methods in classdocs
Each class doc now has two new sections for inherited methods, which shows
number of methods inherited from each superclass, if any. Methods already
documented in this file are excluded. A [show] toggle expands a list of
these methods, as links to the corresponding docs in their superclass.

Also changed how docmap is exported to JS, now as a dictionary instead
of an array. Also some other cleanups and tweaks..