Qt: QFont: add point size management
This adds the concept of point size to QFont, as opposed to traditional
size which stands for pixel size.
Pixel size is system-independent size in pixels, while point size depends
on the system and user's preferences.
Point size can be expressed with a floating point number, while pixel
size needs to be an integral number and will be rounded to nearest integer
otherwise.

New methods are added to the QFont class:
- 'pixelSize'_ and 'pointSize_' setters set the size returned by
  'size' and the boolean returned by 'hasPointSize'
- 'pixelSize' and 'pointSize' getters return the size or nil if
  'hasPointSize' does not match the type of size asked for.

The QFont constructor takes an extra 'usePointSize' boolean argument that
defines how to interpret the given size (the argument defaults to false).