sclang: class tree - sort via std::set
sorting the classes via insertion-sort is rather slow (O(n^2)). to avoid
this, we fill a std::set (a binary search tree) with the class pointers
and traverse this tree to build the sorted list.

this reduces the compilation time of the class library by about 150ms on my
core i7 workstation.

Signed-off-by: Tim Blechmann <tim@klingt.org>