class library: SynthDescLib - upload all synthdefs to server after booting
when synths are added to the global SynthDescLib, they are not uploaded
to servers automatically. with this patch all synthdefs of the global
SynthDescLib are uploaded once a server has successfully booted.

therefore it is not necessary any more to call SynthDef-storeOnce to define
synthdefs before starting a server, but one can simply use SynthDef-add.

Signed-off-by: Tim Blechmann <tim@klingt.org>