supernova: server scheduler - use spsc callback system
the audio sync callback system has only a single consumer (the main audio
thread), but multiple produces (the system thread and the network receiver
thread. however we can use an spsc callback system if we manually
guard the add_callback call.

this should reduce the load of the main audio thread.

Signed-off-by: Tim Blechmann <tim@klingt.org>