supernova: use padded_rw_spinlocks to guard busses
the spinlocks for guarding the busses are provided as an array. this has the
effect that locks for adjacent busses are likely to be located in the same
cache line. therefore locking adjacent busses can possibly result in
unnecessary contention, which can be avoided if we force adjacent locks to
separate cache lines

Signed-off-by: Tim Blechmann <tim@klingt.org>