Fix PMOsc doc: index -> pmindex
As the Beginner's Tutorial by David Michael Cottle in The SuperCollider
Book puts it:

"This assignment, if completed correctly, will give you an error:
WARNING: keyword arg 'index' not found in call to Meta_PMOsc:ar. The
Help file says the third argument is index, but that's incorrect. With
all the advantages of an open source project, there is a flaw -- changes
(this keyword was changed) don't always make it into the documentation."

I do not desire to be uncourteous toward David Michael Cottle (I'm
learning SC through his writings), but I'm quite troubled by the fact
that he has bothered to write this paragraph, but not to fix the
problem.

Signed-off-by: Noe Rubinstein <noe.rubinstein@gmail.com>