HelpBrowser: use SCDoc.findHelpFile for search box and rename to "Quick lookup"
This allows one to go directly to a specific helpfile by entering
a help url (like "Reference/Literals") or the Methods overview
by entering a methodname, or class reference by entering a class name.
Everything else opens in the Search page. If the string is prefixed
with '#' it always opens the Search page with the string.