Close and clear sndfile handle in /b_free command handler.
Free'ing a buffer means that we're done with it, so it should not keep
any sndfile handle open.

This fixes a bug where a new buffer would be unusable by cueSoundFile
if the buffer index happened to have been used before with cueSoundFile
and the user forgot to /b_close it.