Don't resolve symlinks in String-standardizePath. Add new String-realPath that does.
This makes String-standardizePath, Class-filenameSymbol and String-realPath behave
the same on all platforms regarding symlinks.

String-standardizePath did never resolve symlinks on OSX, only in linux.
Resolving symlinks means using realpath() which would always return an absolute path,
but a couple of uses in main class lib expects it not to do that.
The new String-realPath works the same on all platforms that supports symlinks, and
returns nil for non-existent paths.