Fix Ppar and many other patterns that change ~delta in the output event
SimpleNumbers are supposed to be valid as rests, but they break Ppar, Pfindur
and several other patterns that change the result event's delta because
those patterns expect to be able to "put" into the event. This solution
is to make those patterns call asEvent on the stream's results. SimpleNumber
then changes itself into Event.silent(this), which the parent pattern can
then modify at will.