Fix pfset: cleanup function did not fire when pattern was stopped manually
cleanup.addFunction was called accidentally on the input event,
not the result event

also I am reversing 'inevent' and 'event' -- this class, stupidly, used
'event' for the input event and 'inevent' for the result (output) event.

This misuse is present in other pattern classes. I don't have time to
review now, but this really needs to be cleaned up. It's really
confusing.