class library: SynthDef - lazy implementation of removeUGen
the original removeUGen implementation had linear complexity, as both
Array.remove and SynthDef-indexUGens a O(n). to avoid performance penalties
of successive calls, we now use a lazy implementation: instead of removing
the ugen directly, we replace it with nil.
when the graph optimization is finished, we remove all nil entries and re-index
the ugens.

this heavily speeds up the generation of large synthdefs.

Signed-off-by: Tim Blechmann <tim@klingt.org>