OSX: Bundle everything inside the .app, except examples/, README, ChangeLog.txt, etc.
This is the standard way to bundle an .app on OSX, and it fixes the library
references in sclang and scsynth binaries, which now links to the libraries
and frameworks bundled with the .app. Note that since sclang links to the
bundled Qt frameworks, it can not be installed system-wide. But you can install
simple wrapper-scripts into a folder in your PATH.

We also remove the hack in SC_StandAloneInfo_Darwin.cpp to detect standalones
and instead simply pass a SC_STANDALONE definition from cmake.

A MOVED_STUFF.txt file was added in case people wonder where everything has gone,
and README.txt and README_OS_X.txt was updated.