cmake: support user-defined libsndfile dependencies
Setting the SNDFILE_FIND_QUIETLY variable did not do the right job.
It would only silence the warning, but the find_library() call would
still fail if the module was used more than once:

The first time it would not fail as the library variable was set by
the user, but would apparently clear the user status from the variable,
so the second time the check would regard it as a non-user variable,
and fail.