sclang(win): remove a busy-waiting loop under mimicry
The implementation of the input-reading thread on Windows was
mistakingly thought to be sleeping while there is no input. In fact,
it was busy-waiting.

The reason is that the Win API function WaitForMultipleObjects does
not work on anyonymous pipes. Instead of waiting, it always returns
immediately as if the pipe was signalled.

Due to limitations of anonymous pipes on Windows, we have no choice
but to use a blocking read. This also means we have no mechanism for
interrupting the reading thread while blocked. Effectively, when
quitting sclang we have no choice but to let the thread live on and
be terminated automatically when the process terminates.