SCDoc bugfix: handle new helpfiles added after first call to getAllMetaData
If you have the helpbrowser open and add a new helpfile,
you can write the path into the Quick Lookup field to view it.
Before this fix, it would crash since the file was not yet in the doc_map cache.
Now, this case is detected and doc_map rebuilt when trying to show a new file.