sc ide: PostWindow: make behavior of Auto Scroll action more intelligible
The action will only change checked state on direct user interaction.
When it is checked, the view will scroll on new posts **as long as it was
already scrolled to the bottom**.
When it is not checked, the view will never scroll automatically.

Moreover, improve the tooltips of various PostWindow actions.