scide: PostWindow - remove unnecessary '\n' character handling
Should fix issue #533, while avoiding regression to issue #436.

The special handling of '\n' was there in the first place because
the SCProcess::scPost(QString) signal was being abused for internal
IDE status notifications. Now that we have a proper status bar, we
can get rid of this complication.

The handling of '\n' was faulty anyway, because on Windows a new line
consists of 2 characters, not jut '\n', so if this is to be reinstated
for some reason, we need a different implementation anyway.