class library: jitlib - Avoiding sync problems with free/play
When NodeProxy free and play was called in immediate sequence, its group failed to be created properly. This patch fixes #612 and addresses the following issues:

Freeing a NodeProxy or Monitor group should use a time stamp
NodeProxy and Monitor should allow the old group to be freed independently from the new group.

The patch implements an optional argument for Node:free to send a bundle with a timestamp.