SCClassLibrary - MIDIOut - fix method newByName for linux
NewByName was not working correctly for Linux. If we want to connect to a MIDI port that we opened ourselves, the port is usually not available. If the found port is of a higher index than the number of ports we opened, we connect to it via the first port that SC has open.

Signed-off-by: Dan Stowell <danstowell@users.sourceforge.net>