SynthDesc/Def: Slightly rearrange the way SynthDef:add add a desc to a SynthDescLib. Add a dependancy hook to SynthDescLib for when new descs are added
The way this happened is a bit convoluted, and made adding a hook non-straightforward. This patch maintains the existing methods so as not to break code. It adds a new method to SynthDescLib. Users will be directed to use that to add to a lib, and only use the old methods to add to a dictionary.

Signed-off-by: Scott Wilson <i@scottwilson.ca>