class library: strings now throw an error when multiChannelPerform is called.
The method multiChannelPerform, when called on a SequenceableCollection, expands the arguments and the collection together performs the selector on all tuples. Because strings do not behave like Arrays with regards to flop (and do not break up in their constituent characters), they behave like a normal object.
Now when a method on a collection is expanded by multiChannelPerform (e.g. clip, linlin, etc.), and the receiver remains unchanged by flop, the selector is performed (using performList in Object:multiChannelPerform) and expanded indefinitely. This cannot be caught directly, because the method that was expanded doesn't know about that it was called from within a multiChannelPerform. Overriding all those methods in String is overly verbose and error prone. Better block this call on strings for now.