sclang: support for receiving nested OSC bundles.
OSC bundles cannot only contain OSC messages, but other bundles, too. This patch adds support to sclang for receiving such nested OSC bundles.

The method recvOSCbundle is not used anymore and was therefore deleted. Message dispatching is solely done via recvOSCmessage.

The function ConvertOSCBundle is not used anywhere, seems to be dead code to me, so it has been removed.