HelpSource: MIDIFunc: show how to free a MIDIFunc
When using several MIDIFunc, assigning all of them to the same variable
will lead to confusing results as each MIDIFunc will continue to exist
on top of the old one.

This commit shows the good practice: add a variable to each MIDIFunc
example line and show how to free them.