scide: Improve ability to distinguish which split is active
The IDE uses a single tab bar for all splits. This means it is often difficult to see which split corresponds to the active tab. This commit provides more visual feedback, indicating which split has focus, and matching the background colour of the number list to those used in active and inactive tabs as appropriate.

Signed-off-by: Scott Wilson <i@scottwilson.ca>