Classlib: Jit: Unique argument names in ScalarSynthControl for NP slots
If you put multiple scalars in different NodeProxy slots, each one
needs a unique control name. This involved some extra surgery
because both the ScalarSynthControl and the scalars'
prepareForProxySynthDef method need to know the slot index,
and this was not previously passed in.

This is possibly controversial, as it changes method signatures.
It is tested and works correctly, however.