class library: node proxy: general refactoring, allow for dynamic bus change
So far, once a node proxy had been initialized, it had to keep the channel-rate combination andadjust its inputs to it. This change allows to change number of channels and rate on the fly as well as make the proxy "elastic" to reshape according to its new source object. All children (other proxies that read from it) and playback (monitor) is adjusted dynamically. By default, this reshaping is off.