class library: appropriately convert control rate signals to audio rate.
When an array of ugens is passed to a ugen like CombL, the rates are insufficiently converted: one audio ugen in the array would stop a number of control rate ugens to be converted, for instance. The method asAudioRateInput is now also implemented for Array to do this conversion properly.

This commit also fixes #1065.