IDE: Workaround to prevent UDP event feedback loop (qt5.3 bug)
After qt5.3, booting the server from the IDE would cause 150%+ cpu usage due to this QT bug:
https://bugreports.qt-project.org/browse/QTBUG-39080

After some fiddling last night, this workaround bypasses god-knows-what mechanisms in Qt socket notification to avoid the feedback. Given all the weirdness I saw in the Qt sockets implementation, it might be better to move this to boost long term, but this works for now. :)