MIDI: Fix a crash on mac when initializing midi after a library recompile.
On mac, sclang can crash when MIDI is initialized after a library recompile (on my machine and a few others I've used, this occurs 100% of the time). This crash occurs in prListMIDIEndpoints, due to a null devname or endname because of a previous call that errors out. Trivially, we can stop the crash by protecting against cases where MIDIObjectGetStringProperty doesn't return a proper result. These calls were ultimately failing because calling MIDIClientCreate a second time after called MIDIClientDispose results in that call failing. The documentation is unclear exactly how CoreMIDI expects these to be called, but it implies that the normal lifetime of a client is one per app launch ("It is not essential to call this function; the CoreMIDI framework will automatically dispose all MIDIClients when an application terminates.") - in any case, our use case is pretty simple, and a crash is a crash..... This fixes the core problem by only initializing MIDI once, and then keeping that MIDI client for the rest of sclang's lifetime. Effectively, gMIDIClient becomes a singleton.