travis: Ignore homebrew errors.
Homebrew returns non-zero when install / upgrade detects an existing version of the desired package, even if the installed version is perfectly fine (i.e. up to date). There's an open bug on this (https://github.com/Homebrew/homebrew/issues/30939), but until it's dealt with, we'll have to ignore errors from brew install. :(