Create Image.schelp
I translated the SCImage to Image. There are methods that don't work. I left them in there rather than commenting out or deleting, as they should probably be implemented in the Image primitive. Also, it seems to me that jpgs don't work. Not a perfect helpfile but better than nothing and perhaps an encouragement to actually finish the Image primitive.