lang: Ensure sclang works even if stdin isn't available
When launched via double-clicking on it's app bundle, sclang throws an exception when trying to connect to stdin. This fixes. It's not clear exactly why stdin isn't available - but in case of launching via Finder, there would be nothing to read from stdin anyway, so this should have no bearing.