class library: env release trick not needed anymore.
This is the only potentially intrusive fix in this branch - EnvGate could, since it had the Impulse.ar(0) added to the gate input, be released with gate = 0 even in the same block the synth was created. This was added a while ago to avoid hanging synths in node proxy. Now it needs to be released with a gate <= -1 in the first block, and the node proxy does this.

On the other hand, it can now be started with gate = 0 and started later.