class library: protect demand ugens from sum3 and sum4 optimization
The optimization of sums with three or four arguments to Sum3 and Sum4 isn't applicable to demand rate ugens and causes silent hard to track errors. On the long run, a demand rate implementation of Sum3 and Sum4 is preferrable, but as a bugfix it is better to prevent optimization.