scide: add scdoc support
This patch allows scide to call scdoc functions such as scdoc_parse_file(). It
adds a scdoc_log.cpp file to implement error() and post() functions needed by
scdoc.

The aim of this patch is to prepare the autocompleter help: the display of
additionnal help (description, summary, examples) when completing classes.