classlib: refactor quarks save/load
Quarks.load was set up to parse paths both in local=url@refspec form and quark@refspec form, but the latter was never happening because of an extra ? in the regexp. This is fixed, and the parsing broken out into a separate function.

In addition, this change refactors Quarks.save slightly - in cases where we are installing (a) known quarks from the directory, (b) to the default quarks install location, (c) and they are not locally modified, we simply save using quarkname@refspec. The path is platform specific (even possibly machine specific), and is redundant if they're in a default location. For the 99% case (quarks installed from the directory) this makes for a much cleaner list, that is much more likely to be valid across platforms / machines / configurations, and through future changes to SuperCollider.