classlib: introduce QuarkDependency class
This class represents a dependency on some version of a quark, though it may or may not have an exact version. isMetBy reports whether a name@version combination is enough to meet a dependency. conflictsWith reports whether a name@version combination directly conflicts with a dependency. Since tags are generally used as quark versions, versions specified in the form "tags/1.2.3" can be compared transparently with "1.2.3".