classlib: check for dependency conflicts
Once we have a list of quarks back from deepDependencies, make sure there are no conflicts with the list of all dependencies. This would be cases where A -> [B, C@1] and B -> [C@2]. Previously, we could end up with whatever we resolved last (C@2), and a conflict with A's requirement for C@1. Now, we bail and warn.