classlib: fix case differences in quark names
Previously, quarks whose names were specified with differing cases (i.e. "Arduino" and "arduino") would behave MOSTLY as different quarks. Quarks are stored by name and must support case-insensitive filesystems, so quark names in turn must be case-insensitive - so, we need to treat quarks with differing cases as the SAME. BUT -- we need to avoid CHANGING the quark name as it is specified, otherwise it might result in the same quark being installed in two places on case-SENSITIVE filesystems. Ugh. This changes quark name comparisons to be case-insensitive in all cases, and fixes several bugs related to duplicate quarks being installed. In addition, it routes all quark name comparison through a single function, so we can change easily in the future. Finally, when you create a quark via Quark("name"), we try to make that name match whatever is listed in the directory. This should help avoid other name comparison problems, and weird aesthetic differences (i.e. seto versus SETO quark depending on who installed...).