qtcollider: defer setPropery if called inside drawFunc
This should avert a crash when setting the background color of a UserView from inside a drawFunc. In general, it seems extremely unsafe to synchronously set a property of a UserView while *inside* of it's drawFunc. Most properties will behave more or less identically if the set is deferred, and for some (i.e. changing view geometry) this avoids very undefined and dangerous-seeming behavior.