ide: fix crash on session switch
The prevents a crash when iteratively closing all documents. Each doc close triggers a chain of signals that can result in - what appears to be - an already-removed document being referenced by a tab. This change prevents the crash directly by disconnecting the currentChanged signal, which avoids the chain of signals. It also adds some seatbelts in case of bad / null documents, and uses deleteLater() to delete the document, rather than a raw delete. The Document*'s should probably be changed to smart pointers - the signal logic when managing documents is complex enough that it's hard to analytically show that there isn't some pending signal or reference to a document floating around after DocumentManager::close(*) is called.