lang: NetAddr: correctly disconnect tcp socket
NetAddr#disconnect should close the underlying boost TCP socket in order for the port to be available again for use, and should handle the common error that happen when shuting down the TCP connection, if the connection was already terminated from the other side. After the SC_TcpClientPort is closed it is no longer usable therefore the socket variable of NetAddr should be set to null, making NetAddr#isConnected correctly report the status of the connection.