sclang: workaround webkit scrollbar hiding bug
A bug in qt webkit results in animations related to scrollbar opacity and hidden-ness pointing to the QWebView instead of the scrollbars (see https://bugs.webkit.org/show_bug.cgi?id=104116). As a result, the QWebView gets hidden instead of the scrollbars in many cases. This bug was seemingly fixed, but in a very specific and limited way - either it has regressed, or we're running afoul because of our custom style. Regardless, we can work around by implementing the same fix referenced in the aforementioned bug in our own custom styles. Also  - this change removes an OLD webkit scrollbar hack, because it appears to no longer be a problem. This MAY fix the problems with help disappearing in the IDE as well.