sclang: fix crash during SerialPort cleanup
The SerialPort destructor calls join on it's thread, which crashes if run from the thread itself. To close, SerialPort is flagged such that the thread will exit it's run loop, close the port, and then call prDoneAction, which in turn calls prCleanup - this ALWAYS happens from the SerialPort thread itself, meaning it always crashes.

This fixes by (a) ensuring that prCleanup always defers to the main thread, (b) erroring the primitive if we try to call it from the SerialPort thread, (c) asserting if we hit the destructor on the SerialPort thread (which should make problems here easier to track, at least). We are currently not destroying SerialPort backend objects (or removing the threads) properly when the language quits - I'll fix this in a separate commit, though there are no bad effects at the moment (unless you count computer scientists giving annoyed looks).