ide: Workaround for completion help view crash
Dragging or copying text from the help view that pops up during completion results in a crash. It appears that QT deletes the QMimeData corresponding to the copied/dragged chunk too soon, and referencing it later on causes a crash. I attempted various overrides of createMimeDataFromSelection (returning e.g. plain text, etc), and it always crashed except when we simply return NULL. This results in text from that view not being copyable / draggable but this is better than a crash. This is probably a deep QT bug.