Update Ringz.schelp
I made this change because the help file, as it was originally written, did really throw me off by believe that both Resonz and Ringz were in fact the exact same filter, being the only exception how you set the filter parameters. I added more information regarding the relationship between Resonz and Ringz so the differences and similirarities between them is made clearer. 

The help file of Resonz should also be updated to also state the relatioship between these two objects. It doesn't make sense to make the relationship in this help file but completely omit it in the other one. So I'm also going to propose some changes in the help file of Resonz to make it like this one.