Update Resonz.schelp
I made this changes because the help file of Ringz did relate to this object, but the opposite didn't happen. I think it only makes sense to point the relationship in both help files.

By the way, I did also just proposed a change in Rigngz's help file, as the relationship to Resonz was misleading.