Supernova: Don't print spurious late messages for immediate bundles
Recent changes (I think from the timedll branch) cause supernova
to print a "late" message whenever it receives a bundle with an
"immediate" timestamp. This PR eliminates the unnecessary posts,
by adding a member function to time_tag that returns true if the
time tag is immediate, and then using it in
sc_osc_handler::handle_bundle.