class library and ide: display unresponsive server app.
Some refactoring of the serverRunning_ method. This method is now really used to
set the state if known. If unknown, like from the aliveThread, we call
updateRunningState instead. This method will not change serverRunning if false,
but will mark the server as unresponsive. Only if we quit the server from sclang
explicitly, we call serverRunning_ directly.