plugins: GVerb: fix roomsize logic in Ctor
Setting the roomsize to < 1.0 in Ctor triggers unwanted constant
volume noise:

// this first
(
SynthDef(\buggy_GVerb, {
	var wave = GVerb.ar(
		SinOsc.ar * EnvGen.kr(Env.perc(0.1,1)),roomsize:0.1);
	DetectSilence.ar(wave, doneAction: 2);
	Out.ar(0,wave);
}.add;
)

// then spawn a few of these
Synth(\buggy_GVerb)

This behavior is already prevented in the _next() function by
requiring roomsize >= 1.0 and roomsize <= (maxroomsize - 1). This
patch adds the same logic to the constructor. Also, to be consistent
the maxroomsize now has a minimum value of 2 in the constructor.