class library: server - correct limits for allocators
So far, the allocator sizes were calculated by adding the offset to the total number (e.g. numControl + controlBusOffset). But the server received only the total number as initialisation size. This may lead to numbers that the server doesn't recognise. This fix uses the offset only as an offset within the range of allocated numbers instead.