server: Assume requested SR was successfully set
Setting sample rate in CoreAudio is actually an asynchronous operation - you request a sample rate change, and some time in the future it (probably) happens and you are called back about it. This is a problem for us because (a) SC doesn't support the sample rate changes once the server is started, and (b) our startup routine is totally synchronous anyway. Prior to this change, when we requested a custom sample rate, we would do the SetProperty and then shortly after ask CoreAudio for the device sample rate. If the operation had not been completed yet, we would get the OLD rate - the device would it's SR to the requested shortly after, and we would effectively be running the server at the wrong rate. Now, we *assume* that a successful SetProperty has changed the sample rate, and just use the rate we set instead of requesting from CoreAudio. We can still be incorrect if the SR doesn't actually change, but at least we're correct MOST of the time. To increase the changes of success, we query CoreAudio for valid sample rates, and set our requested rate based on those.

A proper fix for this would be to have a properly asynchronous driver startup procedure, where we request all the properties we want, and then wait for callbacks indicating the changes have "taken" before continuing. Note that requesting a buffer size still has the same problem - the fix should be identical - this should be fixed in a different CL.