plugins: fix EnvGen_next_aa
In 730070ad4a29b319dc05bdd3ec26d493eda3ee21, doneAction 14 was fixed, but it broke EnvGen_next_aa:

{ SinOsc.ar(440, 0, EnvGen.ar(Env([0, 1, 0],[0, 0.5], -3), Impulse.ar(1))) }.play

In order to fix this now quickly, this patch comments out the code that activates both the fix and the new bug. The rest is left in place, because it seems a reasonable refactoring that can be used later to make it work properly.