class library: unordered concatenation
Because wslib is very popular and implements ++ (concatenation) for dictionary, there is a lot of code around that breaks when wslib is removed. While usually, concatenation assumes ordered collections, it does make sense for unordered ones and is less error prone than writing:

collection.copy.addAll(otherCollection)

and for dictionary:

collection.copy.putAll(otherCollection)

So the best solution is to add ++ also for unordered collections, which is what this commit does. Fixes #1653.