Improving on correctness of First Steps section in SC Tutorial
As discussed under issue
https://github.com/supercollider/supercollider/issues/1781#issuecomment-
170959978 I have now made an attempt to improve on the correctness of
the section "02. First Steps" of Tutorial "Getting Started With
SuperCollider”. This should give new users of SC a much more pleasant
experience and avoid the bad impression of SC not working as it
should/described. However, some sections are related to a bug
(https://github.com/supercollider/supercollider/issues/1786).