readme: add a build status badge.
I really hate having a chunk of weird URL code a the top of a README.md. But, I am always blown away by the number of major OS projects on GitHub that constantly have failing builds. SuperCollider is honestly pretty rock-solid when it comes to build failures, so we should display our green badge with pride.