lang: detach the resinc thread before exiting
Following the problem of the error code 3 when exiting sclang, it looked
like a problem releasing a thread. So I am detaching it here.
Perhaps it would be better to have a mechanism to signal the thread in
PyrSched, so that it exists.

Now the exit code is 0, like it should be, and I do not see any leaks,
at least under Windows.