midi: sysex output on windows
First stab at making the SysEx output work on Windows. Note that I have
not tested it with real HW yet, but I can see it getting to the PortMIDI
function to send the data, so it should be ok.

I am not sure if I have the proper external equipment to test it
properly, that's why I am publishing it now.