scsynth: line option to bind to specific address (udp)
Add a new command option (-B) to bind to an specific address. The idea
is to allow to bind to the loopback port so that it does not ask for
firewall permissions.

The default is 0.0.0.0, to not change the current behaviour.

This PR is for UDP only. TCP to follow.