class library: refactor nextTimeOnGrid
This applies the binary op dispatch scheme to the commutative
`nextTimeOnGrid`: Because both `clock.nextTimeOnGrid(number)` and
`number. nextTimeOnGrid(clock)` are supposed to be valid, we need one
extra level of dispatch, as to avoid a loop when e.g. `1.
nextTimeOnGrid(1)` is called accidentally.

This is an API change. All clocks must implement `getNextTimeOnGrid`.