sclang: Generic QObject-descendant TypeCodec
This TypeCodec can be used to convert read and write any QObject-derived class. This allows hooking directly into meta-calls that take QObject-derived types as arguments.

In the read case, it checks to see if the slot contains a QObjectProxy. If so, it attempts to qobject_cast it to the requested type, else it provides a null pointer. Note that read == safeRead: this will always provide a null pointer in cases where the type conversion doesn't work.

In the write case, we simply do the reverse - cast to the expected type, and then convert as a QObject, which will connect with it's QObjectProxy (if it exists).