hacking supercollider css colors
cheers. working on hacking supercollider css colours. my goal was to get everything which is somewhat related to background black, and getting the lettering white, with comments and variable names and class names in grey. i accomplished most of that. but within somehow, the variables names, and the classes are still black. if someone can help me to fix this that would be great. this look works far much better for me. don't know how it is for other people. feel free to drop feedback