Topic/rate fallthrough (#1835)
* class library: assume scalar rate as minimum

When determining the rate of a collection of objects, the maximal rate
is returned. Now if the rate happens to be nil, we just assume scalar.
This allows the error to be thrown a bit later, in the UGen that takes
the object as an input. This fixes #1771

* class library: use explicit parentheses

… as not to give rise to confusions and assume the function as the
receiver of minItem.

* class library: throw error when Out has input nil

Normally, this kind of error is caught in `checkValidInputs`. But as
the args are appended, and nil just appends itself as nothing, we never
see this as an input - it simply is missing. This commit explicitly
checks the input size.