lang: remove several debug messages on startup
In an effort to clean up the output that sclang produces in the post window on boot, this commit removes several extraneous debug messages. In particular:

- initPassOne started/done
- NumPrimitives = #
- pass 1 done
- numentries = # / # = #
- # method selectors, # classes
- method table size # bytes, big table size #
- Number of Symbols #
- Byte Code Size #
- compiled # files in # seconds

None of these messages are useful to anyone except SuperCollider developers, so they have been removed. Care must be taken in removing further messages, because supercollider.js (and possibly other interfaces to sclang?) parses the startup output.